refactor: UserNotFoundException user 폴더 하위로 이동
