+++ apiserver/common/src/main/java/com/zipline/global/exception/custom/user/UserNotFoundException.java
@@ -1,7 +1,8 @@
-package com.zipline.global.exception.custom;
+package com.zipline.global.exception.custom.user;
+
+import org.springframework.http.HttpStatus;
 
 import com.zipline.global.exception.BaseException;
-import org.springframework.http.HttpStatus;
 
 public class UserNotFoundException extends BaseException {
 	public UserNotFoundException(String message, HttpStatus status) {
