+++ apiserver/service/src/main/java/com/zipline/service/agentProperty/AgentPropertyServiceImpl.java
@@ -12,9 +12,9 @@ import com.zipline.entity.agentProperty.AgentProperty;
 import com.zipline.entity.customer.Customer;
 import com.zipline.entity.user.User;
 import com.zipline.global.exception.custom.PermissionDeniedException;
-import com.zipline.global.exception.custom.UserNotFoundException;
 import com.zipline.global.exception.custom.agentProperty.PropertyNotFoundException;
 import com.zipline.global.exception.custom.customer.CustomerNotFoundException;
+import com.zipline.global.exception.custom.user.UserNotFoundException;
 import com.zipline.global.request.PageRequestDTO;
 import com.zipline.repository.agentProperty.AgentPropertyRepository;
 import com.zipline.repository.customer.CustomerRepository;
