+++ apiserver/service/src/main/java/com/zipline/service/contract/ContractServiceImpl.java
@@ -17,9 +17,9 @@ import com.zipline.entity.customer.Customer;
 import com.zipline.entity.user.User;
 import com.zipline.global.config.S3Folder;
 import com.zipline.global.exception.custom.PermissionDeniedException;
-import com.zipline.global.exception.custom.UserNotFoundException;
 import com.zipline.global.exception.custom.contract.ContractNotFoundException;
 import com.zipline.global.exception.custom.customer.CustomerNotFoundException;
+import com.zipline.global.exception.custom.user.UserNotFoundException;
 import com.zipline.global.request.PageRequestDTO;
 import com.zipline.global.util.S3FileUploader;
 import com.zipline.repository.contract.ContractDocumentRepository;
