+++ apiserver/service/src/main/java/com/zipline/service/counsel/CounselServiceImpl.java
@@ -16,8 +16,8 @@ import com.zipline.entity.customer.Customer;
 import com.zipline.entity.user.User;
 import com.zipline.global.exception.custom.CounselNotFoundException;
 import com.zipline.global.exception.custom.PermissionDeniedException;
-import com.zipline.global.exception.custom.UserNotFoundException;
 import com.zipline.global.exception.custom.customer.CustomerNotFoundException;
+import com.zipline.global.exception.custom.user.UserNotFoundException;
 import com.zipline.global.response.ApiResponse;
 import com.zipline.repository.counsel.CounselDetailRepository;
 import com.zipline.repository.counsel.CounselRepository;
