+++ apiserver/service/src/main/java/com/zipline/service/customer/CustomerServiceImpl.java
@@ -14,8 +14,8 @@ import com.zipline.entity.counsel.Counsel;
 import com.zipline.entity.customer.Customer;
 import com.zipline.entity.user.User;
 import com.zipline.global.exception.custom.PermissionDeniedException;
-import com.zipline.global.exception.custom.UserNotFoundException;
 import com.zipline.global.exception.custom.customer.CustomerNotFoundException;
+import com.zipline.global.exception.custom.user.UserNotFoundException;
 import com.zipline.global.request.PageRequestDTO;
 import com.zipline.global.response.ApiResponse;
 import com.zipline.repository.agentProperty.AgentPropertyRepository;
